+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `WebhookAttempt.webhookEvent` property
 - Add `AccountSettings.nonRenderableTags` property
 - Add `query` parameter to `EnvelopeRecipientsEndpoint.update` method
+- Add `BillingProfile` endpoint and resource
 
 ## [2.11.0] - 2025-11-12
 ### Deprecated